modularize point interactivity
