+++ src/callbacks/onDraw.js
@@ -1,6 +1,6 @@
 import { updateQuadrantData } from './onDraw/updateQuadrantData';
 import { setDomain } from './onDraw/setDomain';
-import { clearParticipantDetails } from './onResize/clearParticipantDetails';
+import { clearParticipantDetails } from './onResize/customizeMaxPoints/clearParticipantDetails';
 import { updateFilterLabel } from './onDraw/updateFilterLabel';
 import setCutpointMinimums from './onDraw/setCutpointMinimums';
 import syncCutpoints from './onDraw/syncCutpoints';
@@ -8,7 +8,7 @@ import hideEmptyChart from './onDraw/hideEmptyChart';
 import updateStudyDayControl from './onDraw/updateStudyDayControl';
 
 export default function onDraw() {
-    //clear participant Details
+    //clear participant Details (if they exist)
     clearParticipantDetails.call(this);
 
     //get correct cutpoint for the current view
