+++ src/callbacks/onResize/customizeMaxPoints/addPointClick.js
