+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/addExtraMeasureToggle.js
