+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/addFootnote.js
