+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/drawMeasureTable.js
