+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/hideMeasureTable.js
