+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/defaultSettings.js
