+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/init.js
