+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/onDraw/setDomain.js
