+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/onResize/addPointTitles.js
