+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/onResize/drawNormalRange.js
