+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/onResize/drawPopulationExtent.js
