+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/onResize/updatePointFill.js
