+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/lineChart/util/insertAfter.js
