+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/makeNestedData.js
