+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/sparkLine/addSparkClick.js
