+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/measureTable/sparkLine/addSparkLines.js
