+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/participantHeader/clearParticipantHeader.js
