+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/participantHeader/makeParticipantHeader.js
