+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/init.js
