+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/onDraw.js
