+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/onLayout.js
