+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/onPreprocess.js
