+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/onResize.js
