+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/onResize/addExposure.js
