+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/onResize/addPointTitles.js
