+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/onResize/drawCutLine.js
