+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/settings/controlInputs.js
