+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/spaghettiPlot/settings/defaultSettings.js
