+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/visitPath/clearVisitPath.js
