+++ src/callbacks/onResize/customizeMaxPoints/addPointClick/visitPath/drawVisitPath.js
