+++ src/callbacks/onResize/customizeMaxPoints/addPointMouseover/clearRugs.js
