+++ src/callbacks/onResize/customizeMaxPoints/addPointMouseover/drawRugs.js
