+++ src/callbacks/onResize/customizeMaxPoints/addPointTitles.js
