+++ src/callbacks/onResize/customizeMaxPoints/clearParticipantDetails.js
@@ -7,9 +7,13 @@ import { formatPoints } from './formatPoints';
 export function clearParticipantDetails() {
     var chart = this;
     var config = this.config;
-    var points = this.svg.selectAll('g.point').select('circle');
+    var points = this.svg
+        .select('g.point-supergroup.mark1')
+        .selectAll('g.point')
+        .select('circle');
 
     points.classed('disabled', false);
+
     this.config.quadrants.table.wrap.style('display', null);
     clearVisitPath.call(this); //remove path
     clearParticipantHeader.call(this);
@@ -17,5 +21,6 @@ export function clearParticipantDetails() {
     clearRugs.call(this, 'y');
     hideMeasureTable.call(this); //remove the detail table
     formatPoints.call(this);
+
     this.participantDetails.wrap.selectAll('*').style('display', 'none');
 }
