+++ src/callbacks/onResize/customizeMaxPoints/updateParticipantMarks.js
